requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
numpy>=1.25.0
Pillow>=10.0.0
scikit-learn>=1.3.0

